--- Bccupass_CoreMVC/Common/Enum/OrderState.cs
@@ -1,16 +1,0 @@
-﻿using System.ComponentModel;
-
-namespace Bccupass_CoreMVC.Common.Enum
-{
-        public enum OrderState
-        {
-            [Description("未付款")]
-            NotPaid,
-            [Description("已付款")]
-            Paid,
-            [Description("已取消")]
-            Cancel,
-            [Description("已退票")]
-            Refund
-        }
-}
