+++ Bccupass_CoreMVC/Common/Enums/ActivityState.cs
