+++ Bccupass_CoreMVC/Common/Enums/ActivityStateByTime.cs
