+++ Bccupass_CoreMVC/Common/Enums/ActivityTheme.cs
