+++ Bccupass_CoreMVC/Controllers/UserController.cs
@@ -1,4 +1,5 @@
-﻿using Bccupass_CoreMVC.Models.DTO.Ticket;
+﻿using Bccupass_CoreMVC.Common.Enum;
+using Bccupass_CoreMVC.Models.DTO.Ticket;
 using Bccupass_CoreMVC.Models.ViewModel;
 using Bccupass_CoreMVC.Models.ViewModel.User;
 using Bccupass_CoreMVC.Services;
@@ -23,31 +24,13 @@ namespace Bccupass_CoreMVC.Controllers
             return View();
         }
 
-        public IActionResult UserTicket(int userId)
+        public IActionResult UserTicket(int userId, int selcetByOrderState,int pageId=1)
         {
-
-            //int activePage = 1; // 目前在第幾頁
-            //int pageRows = 5; // 一頁幾筆資料
-
+            //userId = int.Parse(User.Identity?.Name);
             userId = 1;
             var userTicketDto = _userService.GetOrderListOfUser(userId);
             var order = userTicketDto.Select(x => _ticketService.GetTicket(x.OrderId));
             var total = userTicketDto.Count();
-
-            //// 計算頁數
-            //int pages = 0;
-            //if (total % pageRows == 0)
-            //{
-            //    pages = total / pageRows;
-            //}
-            //else
-            //{
-            //    pages = (total / pageRows) + 1;
-            //}
-
-            //int startRow = (activePage - 1) * pageRows;
-
-
 
             var orderList = order.Select(x =>
                  new CreateTicketViewModel()
@@ -85,45 +68,28 @@ namespace Bccupass_CoreMVC.Controllers
                      TicketNum = _ticketService.TicketCount(x.Order.OrderId)
                  });
 
+            switch (selcetByOrderState)
+            {
+                case (int)OrderState.Paid:
+                    orderList = orderList.Where(x=>x.Order.OrderState == selcetByOrderState).ToList();
+                    break;
+                case (int)OrderState.NotPaid:
+                    orderList = orderList.Where(x => x.Order.OrderState == selcetByOrderState).ToList();
+                    break;
+                case (int)OrderState.Cancel:
+                    orderList = orderList.Where(x => x.Order.OrderState == selcetByOrderState).ToList();
+                    break;
+                case (int)OrderState.Refund:
+                    orderList = orderList.Where(x => x.Order.OrderState == selcetByOrderState).ToList();
+                    break;
+            }
+
             var res = new UserTicketViewModel()
             {
                 OrderList = orderList
             };
 
             return View(res);
-
-            //var createTicketVM = new CreateTicketViewModel()
-            //{
-            //    User = new UserTicketViewModel.UserData { UserId = userTicketDto.User.UserId },
-            //    Order = userTicketDto.Order.Select(x => new UserTicketViewModel.OrderData()
-            //    {
-            //        OrderId = x.OrderId,
-            //    }),
-            //    TdOd = userTicketDto.TdOd.Select(x => new UserTicketViewModel.TicketDetailOrderDetail()
-            //    {
-            //        TdOdId = x.TdOdId,
-            //        BuyerName = x.BuyerName,
-            //        BuyerCompanyName = x.BuyerCompanyName,
-            //        BuyerPhone = x.BuyerPhone,
-            //        BuyerEmail = x.BuyerEmail
-            //    }),
-            //    TicketDetail = userTicketDto.TicketDetail.Select(x => new UserTicketViewModel.TicketDatail()
-            //    {
-            //        TicketId = x.TicketId,
-            //        TicketName = x.TicketName,
-            //        Price = x.Price
-            //    }),
-            //    Activity = userTicketDto.Activity.Select(x => new UserTicketViewModel.ActivityData()
-            //    {
-            //        ActId = x.ActId,
-            //        ActName = x.ActName,
-            //        StartTime = x.StartTime,
-            //        EndTime = x.EndTime
-            //    })
-            //};
-
-            //return View(userTicketVM);
-
 
         }
