+++ Bccupass_CoreMVC/Models/ViewModel/User/UserTicketViewModel.cs
@@ -6,6 +6,9 @@ namespace Bccupass_CoreMVC.Models.ViewModel.User
     public class UserTicketViewModel
     {
         public IEnumerable<CreateTicketViewModel> OrderList { get; set; }
+        public int selcetByOrderState { get; set; }
+
+
     }
 
 }
