+++ Bccupass_CoreMVC/Views/Shared/_userTicketPartial.cshtml
@@ -13,7 +13,7 @@
                         <p class="id fs-6 my-2 text-secondary">@Model.Order.OrderId</p>
                     </div>
                     <div class="ticket-operate col-12 col-md-4 d-flex align-items-center ">
-                        <a class="get btn btn-primary border me-3  ms-auto d-flex align-items-center justify-content-center p-0 py-3" href="/CreateTicket/@Model.Order.OrderId">快速取票</a>
+                        <a class="get btn btn-primary border me-3  ms-auto d-flex align-items-center justify-content-center p-0 py-3" href="/Ticket/CreateTicket/@Model.Order.OrderId">快速取票</a>
                         <div class="price  btn-white border  me-3  d-flex align-items-center align-content-center flex-wrap justify-content-center p-0 py-3">
                             @Model.TicketDetail.Sum(x=>x.Price)
                             <div class=" text-secondary col-12 text-center">訂單金額</div>
@@ -30,7 +30,7 @@
                         </div>
                     </div>
                 </div>
-                <div id="panelsStayOpen-collapseOne" class="accordion-collapse collapse show"
+                <div id="panelsStayOpen-collapseOne" class="accordion-collapse collapse"
                      aria-labelledby="panelsStayOpen-headingOne">
                     <table class="table fs-6">
                         <thead>
