+++ Bccupass_CoreMVC/Views/User/UserTicket.cshtml
@@ -1,8 +1,9 @@
-﻿@using Bccupass_CoreMVC.Models.ViewModel.User
+﻿@using Bccupass_CoreMVC.Common.Enum
+@using Bccupass_CoreMVC.Models.ViewModel.User
 @model UserTicketViewModel
 
 @section Topcss {
-<link rel="stylesheet" href="~/css/Ticket/CreateTicket.css">
+<link rel="stylesheet" href="~/css/User/UserTicket.css">
 }
 
 @section TopJS{
@@ -14,17 +15,17 @@
     <div class="ticket-nav col-lg-12 col-md-12 col row mx-3 my-1 py-3">
         <h1 class="fs-4 fw-bolder ">我的票券</h1>
         <ul class="Nav-ticket nav nav-tabs col-12 ">
-            <li class="nav-item  border border-bottom-0  ">
-                <a class="nav-link active" aria-current="page" href="#">已報名</a>
+            <li class="nav-item  border border-bottom-0 " role="presentation">
+                <a asp-action="UserTicket" asp-route-selcetByOrderState="0"  class="nav-link @(Model.selcetByOrderState == (int)OrderState.Paid? "active":"")" id="orderState-tab" role="tab" aria-controls="orderState" aria-selected="true">已報名</a>
             </li>
-            <li class="nav-item   border border-bottom-0 ">
-                <a class="nav-link text-black" href="#">未完成</a>
+            <li class="nav-item   border border-bottom-0 " role="presentation">
+                <a asp-action="UserTicket" asp-route-selcetByOrderState="1" class="nav-link @(Model.selcetByOrderState == (int)OrderState.NotPaid? "active":"")" id="orderState-tab" role="tab" aria-controls="orderState" aria-selected="false">未完成</a>
             </li>
-            <li class="nav-item text-black  border border-bottom-0">
-                <a class="nav-link text-black" href="#">已取消</a>
+            <li class="nav-item text-black  border border-bottom-0" role="presentation">
+                <a asp-action="UserTicket" asp-route-selcetByOrderState="2" class="nav-link @(Model.selcetByOrderState == (int)OrderState.Cancel? "active":"")" id="orderState-tab" role="tab" aria-controls="orderState" aria-selected="false">已取消</a>
             </li>
-            <li class="nav-item  border border-bottom-0">
-                <a class="nav-link text-black">
+            <li class="nav-item  border border-bottom-0" role="presentation">
+                <a asp-action="UserTicket" asp-route-selcetByOrderState="3" class="nav-link @(Model.selcetByOrderState == (int)OrderState.Refund? "active":"")" id="orderState-tab" role="tab" aria-controls="orderState" aria-selected="false">
                     退票紀錄
                 </a>
             </li>
