+++ Bccupass_CoreMVC/wwwroot/css/User/userticket.scss
@@ -20,7 +20,6 @@
 
 .ticket-operate {
 
-
     .get, .info, .price, .info-btn {
         width: 32%;
         height: 70%;
