+++ Bccupass_CoreMVC/wwwroot/css/home/home.css
@@ -1,2 +1,26 @@
 ﻿.container {
   max-width: 1080px; }
+
+body {
+  background-color: #F6F8F8; }
+
+.activities-list, .create-activity, .everybody-track {
+  max-width: 1080px;
+  margin: 20px auto; }
+
+.create-activity .cta-btn {
+  border-radius: 25px;
+  color: #ffffff;
+  background-color: #2C5E9E;
+  border: 1px solid #2C5E9E; }
+
+.everybody-track .organizer-img {
+  width: 150px;
+  height: 150px;
+  box-shadow: 3px 5px 12px rgba(0, 0, 0, 0.15);
+  border-radius: 50%;
+  overflow: hidden;
+  background-image: url("https://static.accupass.com/org/2103240413414437804150.jpg");
+  background-position: center;
+  background-repeat: no-repeat;
+  background-size: contain; }
